fix(dpa4): address review and CI regressions
